rename get user by id usecase
