+++ organizator_api/tests/users/domain/usecases/test_get_user_by_id_use_case.py
@@ -1,7 +1,7 @@
 from tests.api_tests import ApiTests
 from tests.users.domain.UserFactory import UserFactory
 
-from app.users.domain.use_cases.get_by_id_use_case import GetUserByIdUseCase
+from app.users.domain.use_cases.get_user_by_id_use_case import GetUserByIdUseCase
 
 
 class TestGetUserByIdUseCase(ApiTests):
